Some improvements
Add possibility to use a custom key derivation function
Add Scrypt as key derivation function
A list of salts are required when encrypting, instead of a single one. Salts will be written to the encrypted file.
If not specified, salts are randomly generated instead of using a default one
No need to specify salts when decrypting anymore (they are written in the file)